+++ site-de-notas-futurista/functions/index.js
@@ -172,3 +172,6 @@ exports.scoreExecucaoAltoValorOnWrite = functions.firestore
   .onWrite(scoreExecucaoAltoValorOnWrite)
 
 exports.datajudSearch = datajudSearch
+
+const { syncAfsAdminClaims } = require("./afs-market-admin")
+exports.syncAfsAdminClaims = syncAfsAdminClaims